Finishing validations
Finalizando tratativas de erro e corrigindo validações na entidade ClientDTO.

+++ src/main/java/com/carnacorp/desafio3/services/ClientService.java
@@ -1,6 +1,7 @@
 package com.carnacorp.desafio3.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -9,7 +10,10 @@ import org.springframework.transaction.annotation.Transactional;
 import com.carnacorp.desafio3.dto.ClientDTO;
 import com.carnacorp.desafio3.entities.Client;
 import com.carnacorp.desafio3.repositories.ClientRepository;
+import com.carnacorp.desafio3.services.exceptions.DatabaseException;
 import com.carnacorp.desafio3.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
@@ -19,7 +23,8 @@ public class ClientService {
 
 	@Transactional(readOnly = true)
 	public ClientDTO findById(Long id) {
-		Client client = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado"));
+		Client client = repository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado"));
 		return new ClientDTO(client);
 	}
 
@@ -39,15 +44,26 @@ public class ClientService {
 
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO dto) {
-		Client entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ClientDTO(entity);
+		try {
+			Client entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ClientDTO(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Cliente não encontrado");
+		}
 	}
 
 	@Transactional
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Cliente não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 	}
 
 	private void copyDtoToEntity(ClientDTO dto, Client entity) {
